requirements.txt
# Core AI and LLM dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.12.0

# PDF processing
pypdf>=3.17.0
pdfplumber>=0.10.0

# HTTP and async
requests>=2.31.0
aiohttp>=3.9.0

# Configuration
python-dotenv>=1.0.0

# Vector store and embeddings
faiss-cpu>=1.7.4
tiktoken>=0.5.0

# Academic APIs
arxiv>=2.1.0

# Web scraping and parsing
beautifulsoup4>=4.12.0

# Data validation
pydantic>=2.5.0

# CLI and output formatting
rich>=13.7.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
ruff>=0.2.0
mypy>=1.8.0
